# kisaan-backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: KisaanCenter API
  version: 1.0.0
  description: Full API documentation for KisaanCenter backend
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: integer }
        shop_id: { type: integer }
        farmer_id: { type: integer }
        buyer_id: { type: integer }
        category_id: { type: integer }
        product_name: { type: string }
        quantity: { type: number, minimum: 0 }
        unit_price: { type: number, minimum: 0 }
        total_sale_value: { type: number, minimum: 0 }
        shop_commission: { type: number, minimum: 0 }
        farmer_earning: { type: number, minimum: 0 }
        status: { type: string, enum: [pending, completed, cancelled], default: pending }
        transaction_date: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [shop_id, farmer_id, buyer_id, category_id, product_name, quantity, unit_price]
    TransactionCreate:
      type: object
      properties:
        shop_id: { type: integer }
        farmer_id: { type: integer }
        buyer_id: { type: integer }
        category_id: { type: integer }
        product_name: { type: string }
        quantity: { type: number, minimum: 0 }
        unit_price: { type: number, minimum: 0 }
        transaction_date: { type: string, format: date-time }
      required: [shop_id, farmer_id, buyer_id, category_id, product_name, quantity, unit_price]
    Payment:
      type: object
      properties:
        id: { type: integer }
        transaction_id: { type: integer }
        payer_type: { type: string, enum: [BUYER, SHOP] }
        payee_type: { type: string, enum: [SHOP, FARMER] }
        amount: { type: number, minimum: 0 }
        status: { type: string, enum: [PENDING, PAID, FAILED], default: PENDING }
        payment_date: { type: string, format: date-time, nullable: true }
        method: { type: string, enum: [CASH, BANK, UPI, OTHER], default: CASH }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [transaction_id, payer_type, payee_type, amount]
    PaymentCreate:
      type: object
      properties:
        transaction_id: { type: integer }
        payer_type: { type: string, enum: [BUYER, SHOP] }
        payee_type: { type: string, enum: [SHOP, FARMER] }
        amount: { type: number, minimum: 0 }
        method: { type: string, enum: [CASH, BANK, UPI, OTHER], default: CASH }
        notes: { type: string, nullable: true }
        payment_date: { type: string, format: date-time, nullable: true }
      required: [transaction_id, payer_type, payee_type, amount]
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, default: false }
        message: { type: string }
        error: { type: string }
...
    TransactionSummary:
      type: object
      properties:
        total_transactions: { type: integer }
        total_value: { type: number }
        total_commission: { type: number }
        pending_count: { type: integer }
        completed_count: { type: integer }
        average_transaction_value: { type: number }
...
paths:
...
  /transactions:
    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreate' }
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Transaction' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all transactions
      parameters:
        - in: query
          name: shop_id
          schema: { type: integer }
        - in: query
          name: farmer_id
          schema: { type: integer }
        - in: query
          name: buyer_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [pending, completed, cancelled] }
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Transaction' }
  /transactions/{id}:
    get:
      summary: Get transaction by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Transaction' }
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /transactions/shop/{shopId}:
    get:
      summary: Get transactions by shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop transactions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Transaction' }
  /transactions/shop/{shopId}/earnings:
    get:
      summary: Get shop earnings
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop earnings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TransactionSummary' }
  /transactions/farmer/{farmerId}/earnings:
    get:
      summary: Get farmer earnings
      parameters:
        - in: path
          name: farmerId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Farmer earnings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TransactionSummary' }
  /transactions/buyers/{buyerId}/purchases:
    get:
      summary: Get purchases by buyer
      parameters:
        - in: path
          name: buyerId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Buyer purchases
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Transaction' }
  /payments:
    post:
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentCreate' }
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all payments
      parameters:
        - in: query
          name: transaction_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, PAID, FAILED] }
        - in: query
          name: payer_type
          schema: { type: string, enum: [BUYER, SHOP] }
        - in: query
          name: payee_type
          schema: { type: string, enum: [SHOP, FARMER] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
  /payments/{id}:
    get:
      summary: Get payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
        '404':
          description: Payment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /payments/{id}/status:
    put:
      summary: Update payment status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [PENDING, PAID, FAILED] }
                payment_date: { type: string, format: date-time }
                notes: { type: string }
      responses:
        '200':
          description: Payment status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
  /payments/transaction/{transactionId}:
    get:
      summary: Get payments by transaction
      parameters:
        - in: path
          name: transactionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Transaction payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
...